Generate and Send web push notifications
Generate will create a public and private key pair and then write out a
sample web site using the keys. The sample web site can subscribe to
notifications, at which time the command used to send a notification to
the sample web site will be created. Running this command will send a
notification to the sample web site.